% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Body Mass Index (BMI)}
\usage{
bmi(x, y)
}
\arguments{
\item{x}{Your size in centimeters.}

\item{y}{Your weight in kilograms.}
}
\description{
Return the value of the body mass index (BMI)

BMI = Weight/Size^2
}
\examples{
bmi(x=183, y=75)
}
\author{
Gaëtan Lovey, Cédric Vuignier, Jérémy Bayer, Raffaello Raffin
}
