% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needs.R
\name{needs}
\alias{needs}
\title{Calorie need}
\usage{
needs(bmr, multiplier)
}
\arguments{
\item{bmr}{The value of your basal metabolic rate (bmr).}

\item{multiplier}{Multiplier depending on your weekly physical activities (Very light = 1.30, Light = 1.55, Moderate = 1.65, Heavy = 1.80, Very heavy = 2).}
}
\description{
Return the daily calorie need for a person.
}
\examples{
needs(bmr=1620, multiplier=1.65)
}
