% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking_time.R
\name{walking_time}
\alias{walking_time}
\title{Walking time}
\usage{
walking_time(y, MET, e, f, kcal)
}
\arguments{
\item{y}{Your weight in kilograms.}

\item{MET}{Metabolic Equivalent of Task (normal walking = 3, fast walking = 4.5).}

\item{e}{Fix value of 3.5.}

\item{f}{Fix value of 200.}

\item{kcal}{Calorie intake.}
}
\description{
Returns the walking time in minutes useful for expending the calories absorbed.

Walking time = Kcal/((MET*3.5*Weight)/200)
}
\examples{
walking_time(y=75, MET = 3, e = 3.5, f = 200, kcal = 800)
}
\author{
Gaëtan Lovey, Cédric Vuignier, Jérémy Bayer, Raffaello Raffin
}
